Adjusted CSS Extractors to match Scrapy's vision
I found out that scrapy sees the website a bit differently, in which some of the html code is different. Therefore, I adjusted the css extractors to match what scrapy sees. I found that the normal prices that I saw were not seen by scrapy, instead most items stated that prices were offered from a certain number. These prices weren't the same as what my website shows, so I adjusted for it. For the items without a listed prices, I put "N/A".