Scraped Important Data for Review Scoring. Add Pagination Handling.
I scraped the title, rank, price, imagelink, the number of ratings and the rating score out of 5. The last 2 items were scraped for review scoring, so the items can be later sorted by rating score. I also added pagination handling by scraping all pages for the products, in this case 2 pages for 100 products.