Finished converting, added pagelink
Scraped part info and categories properly. For parts_link, I added the path the page on the local pdf where this part  was listed.  I didn't include the page code from the manual, since most codes don't exist. Going to the page where the data was scraped gives more control and info to user. I tried adding hyperlink, but it doesn't navigate the specific page on local pdfs. However, when I manually put the path in my broswer, the correct page on the pdf opens. The pagelink will only work for local pdfs, so this must be on the user's computer. This can be adjusted for if the pdf is hosted online.